streamlit_app/requirements.txt
streamlit==1.36.0
pandas==2.2.2
psycopg2-binary==2.9.9
plotly==5.22.0
